requirements.txt
django==1.10.3
djangorestframework==3.5.3
django-filter==1.0.0
